Move menus to separate folders
