+++ apps/web/src/components/Menu/Menu.test.tsx
@@ -1,9 +1,9 @@
 import { useColorMode } from "@chakra-ui/system";
 import { downloadBackupFile } from "@umami/state";
 
-import { AddressBookMenu } from "./AddressBookMenu";
-import { AdvancedMenu } from "./AdvancedMenu";
-import { AppsMenu } from "./AppsMenu";
+import { AddressBookMenu } from "./AddressBookMenu/AddressBookMenu";
+import { AdvancedMenu } from "./AdvancedMenu/AdvancedMenu";
+import { AppsMenu } from "./AppsMenu/AppsMenu";
 import { LogoutModal } from "./LogoutModal";
 import { Menu } from "./Menu";
 import {
